(HALF DONE) Partially Converted Imperial Reconquest - MIGHT BE BUGGED - READ DESC (Ancient Work)
So this has been waiting for like a year and a half to get committed, and I think its broken anyway. If it's bugged, try getting rid of the any_title and limit under the can_use_title section, since looking at the code I *think* the OR block alone should be enough to cover it.

This is still only half done.

WHOLE MOD STILL NEEDS TO BE UPDATED TO HOLY FURY/IRON CENTURY.